LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-datalist-element/datalistoptions-expected.txt


FAIL options label/value assert_equals: expected 5 but got 3
PASS If an element has a datalist element ancestor, it is barred from constraint validation 
Number:  
